pyproject.toml
[tool.poetry]
name = "api-django"
version = "0.1.0"
description = ""
authors = ["gabrielemv <gabrielevaladaresm@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.7"
djangorestframework = "^3.15.2"
django-cors-headers = "^4.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
